user and item IDs: support long

+++ src/MyMediaLite/IO/AttributeData.cs
@@ -17,7 +17,6 @@
 
 using System;
 using System.Data;
-using System.Globalization;
 using System.IO;
 using MyMediaLite.Data;
 using MyMediaLite.DataType;
@@ -73,7 +72,7 @@ namespace MyMediaLite.IO
 				if (tokens.Length != 2)
 					throw new IOException("Expected exactly 2 columns: " + line);
 
-				int entity_id = mapping.ToInternalID(int.Parse(tokens[0]));
+				int entity_id = mapping.ToInternalID(long.Parse(tokens[0]));
 				int attr_id   = int.Parse(tokens[1]);
 
 				matrix[entity_id, attr_id] = true;
